+++ app/Widgets/Post/Post.php
@@ -22,9 +22,9 @@ class Post extends Base
         $this->registerEvent('microblog_commentsget_error', 'onCommentsError');
         $this->registerEvent('microblog_commentpublish_handle', 'onCommentPublished');
         $this->registerEvent('microblog_commentpublish_error', 'onCommentPublishError');
-        $this->registerEvent('pubsub_getitem_handle', 'tonHandle', 'post');
         $this->registerEvent('pubsub_postdelete_handle', 'onDelete', 'post');
         $this->registerEvent('pubsub_getitem_errorpresencesubscriptionrequired', 'onPresenceSubscriptionRequired');
+        $this->registerEvent('post', 'tonHandle', 'post');
         $this->registerEvent('post_resolved', 'tonHandle', 'post');
     }
 
@@ -115,7 +115,6 @@ class Post extends Base
         $gi->setTo($server)
             ->setNode($node)
             ->setId($nodeid)
-            ->setManual()
             ->request();
 
         if ($p) {
@@ -133,7 +132,7 @@ class Post extends Base
                 $gi->setTo($p->replyserver)
                     ->setNode($p->replynode)
                     ->setId($p->replynodeid)
-                    ->setAskReply($p->id)
+                    ->setReplypostid($p->id)
                     ->request();
             }
         } else {