+++ public/scripts/movim_tpl.js
@@ -84,6 +84,7 @@ var MovimTpl = {
     },
     remove: function (selector) {
         target = document.querySelector(selector);
+        console.log(selector)
         if (target) {
             target.remove();
         }